+++ packages/core/test/Form.test.jsx
@@ -1,4 +1,4 @@
-import { Component, createRef, useEffect } from 'react';
+import { Component, createRef, useEffect, useState } from 'react';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import { fireEvent, act, render, waitFor } from '@testing-library/react';
@@ -442,10 +442,24 @@ describeRepeated('Form common', (createFormComponent) => {
     });
 
     it('should pass errors as the provided React component', () => {
+      // live validate does not run on initial render anymore
+      expect(node.querySelectorAll('.error-detail li')).to.have.length.of(0);
+      act(() => {
+        fireEvent.change(node.querySelector('input'), {
+          target: { value: 'stillinvalid' },
+        });
+      });
       expect(node.querySelectorAll('.error-detail li')).to.have.length.of(1);
     });
 
     it('should pass rawErrors as an array of strings', () => {
+      // live validate does not run on initial render anymore
+      expect(node.querySelectorAll('.raw-error')).to.have.length.of(0);
+      act(() => {
+        fireEvent.change(node.querySelector('input'), {
+          target: { value: 'stillinvalid' },
+        });
+      });
       expect(node.querySelectorAll('.raw-error')).to.have.length.of(1);
     });
 
@@ -2731,6 +2745,15 @@ describeRepeated('Form common', (createFormComponent) => {
       it('should denote the error in the field', () => {
         const { node } = createFormComponent(formProps);
 
+        // live validate does not run on initial render anymore
+        expect(node.querySelectorAll('.rjsf-field-error')).to.have.length.of(0);
+
+        act(() => {
+          fireEvent.change(node.querySelector('input'), {
+            target: { value: 'shorts' },
+          });
+        });
+
         expect(node.querySelectorAll('.rjsf-field-error')).to.have.length.of(1);
         expect(node.querySelector('.rjsf-field-string .error-detail').textContent).eql(
           'must NOT have fewer than 8 characters',
@@ -2777,6 +2800,14 @@ describeRepeated('Form common', (createFormComponent) => {
       it('should denote the error in the field', () => {
         const { node } = createFormComponent(formProps);
 
+        // live validate does not run on initial render anymore
+        expect(node.querySelectorAll('.rjsf-field-error')).to.have.length.of(0);
+
+        act(() => {
+          fireEvent.change(node.querySelector('input'), {
+            target: { value: 'shorts' },
+          });
+        });
         const liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
         const errors = [].map.call(liNodes, (li) => li.textContent);
 
@@ -2829,6 +2860,14 @@ describeRepeated('Form common', (createFormComponent) => {
 
       it('should denote the error in the field', () => {
         const { node } = createFormComponent(formProps);
+        // live validate does not run on initial render anymore
+        expect(node.querySelectorAll('.rjsf-field-error')).to.have.length.of(0);
+
+        act(() => {
+          fireEvent.change(node.querySelector('input'), {
+            target: { value: 'shorts' },
+          });
+        });
         const errorDetail = node.querySelector('.rjsf-field-object .rjsf-field-string .error-detail');
 
         expect(node.querySelectorAll('.rjsf-field-error')).to.have.length.of(1);
@@ -2848,7 +2887,7 @@ describeRepeated('Form common', (createFormComponent) => {
       const formProps = {
         schema,
         liveValidate: true,
-        formData: ['good', 'bad', 'good'],
+        formData: ['good', 'ba', 'good'],
       };
 
       it('should contextualize the error for array indices', () => {
@@ -2872,10 +2911,19 @@ describeRepeated('Form common', (createFormComponent) => {
         const { node } = createFormComponent(formProps);
         const fieldNodes = node.querySelectorAll('.rjsf-field-string');
 
+        // live validate does not run on initial render anymore
+        expect(fieldNodes[1].classList.contains('rjsf-field-error')).eql(false);
+
+        // Change the End field to a larger value than Start field to remove customValidate raised errors.
+        act(() => {
+          fireEvent.change(fieldNodes[1].querySelector('input'), {
+            target: { value: 'bad' },
+          });
+        });
+
         const liNodes = fieldNodes[1].querySelectorAll('.rjsf-field-string .error-detail li');
         const errors = [].map.call(liNodes, (li) => li.textContent);
 
-        expect(fieldNodes[1].classList.contains('rjsf-field-error')).eql(true);
         expect(errors).eql(['must NOT have fewer than 4 characters']);
       });
 
@@ -2938,13 +2986,25 @@ describeRepeated('Form common', (createFormComponent) => {
         const { node } = createFormComponent({
           ...formProps,
           formData: {
-            level1: ['good', 'bad', 'good'],
+            level1: ['good', 'ba', 'good'],
           },
         });
 
-        const liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
-        const errors = [].map.call(liNodes, (li) => li.textContent);
+        const fields = node.querySelectorAll('.rjsf-field-string');
+        let liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
+        let errors = [].map.call(liNodes, (li) => li.textContent);
 
+        // live validate does not run on initial render anymore
+        expect(errors).eql([]);
+
+        act(() => {
+          fireEvent.change(fields[1].querySelector('input'), {
+            target: { value: 'bad' },
+          });
+        });
+
+        liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
+        errors = [].map.call(liNodes, (li) => li.textContent);
         expect(errors).eql(['must NOT have fewer than 4 characters']);
       });
     });
@@ -3012,14 +3072,28 @@ describeRepeated('Form common', (createFormComponent) => {
 
       it('should denote the error in the nested item field in error', () => {
         const { node } = createFormComponent(formProps);
+
         const fields = node.querySelectorAll('.rjsf-field-string');
-        const errors = [].map.call(fields, (field) => {
+        let errors = [].map.call(fields, (field) => {
+          const li = field.querySelector('.error-detail li');
+          return li && li.textContent;
+        });
+
+        // live validate does not run on initial render anymore
+        expect(errors).eql([null, null, null, null]);
+
+        act(() => {
+          fireEvent.change(fields[0].querySelector('input'), {
+            target: { value: 'bad' },
+          });
+        });
+        errors = [].map.call(fields, (field) => {
           const li = field.querySelector('.error-detail li');
           return li && li.textContent;
         });
 
         expect(errors).eql([
-          null,
+          'must NOT have fewer than 4 characters',
           'must NOT have fewer than 4 characters',
           'must NOT have fewer than 4 characters',
           null,
@@ -3044,7 +3118,7 @@ describeRepeated('Form common', (createFormComponent) => {
       const formProps = {
         schema,
         liveValidate: true,
-        formData: [{ foo: 'good' }, { foo: 'bad' }, { foo: 'good' }],
+        formData: [{ foo: 'good' }, { foo: 'ba' }, { foo: 'good' }],
       };
 
       it('should contextualize the error for array nested items', () => {
@@ -3068,9 +3142,17 @@ describeRepeated('Form common', (createFormComponent) => {
         const { node } = createFormComponent(formProps);
         const fieldNodes = node.querySelectorAll('.rjsf-field-string');
 
-        const liNodes = fieldNodes[1].querySelectorAll('.rjsf-field-string .error-detail li');
-        const errors = [].map.call(liNodes, (li) => li.textContent);
+        // Initial render no longer does live validation
+        expect(fieldNodes[1].classList.contains('rjsf-field-error')).eql(false);
+        // Change the End field to a larger value than Start field to remove customValidate raised errors.
+        act(() => {
+          fireEvent.change(fieldNodes[1].querySelector('input'), {
+            target: { value: 'bad' },
+          });
+        });
 
+        const liNodes = fieldNodes[1].querySelectorAll('.error-detail li');
+        const errors = [].map.call(liNodes, (li) => li.textContent);
         expect(fieldNodes[1].classList.contains('rjsf-field-error')).eql(true);
         expect(errors).eql(['must NOT have fewer than 4 characters']);
       });
@@ -3221,16 +3303,25 @@ describeRepeated('Form common', (createFormComponent) => {
         const { node, onChange } = createFormComponent({
           schema,
           liveValidate: true,
-          formData: { Start: 2, End: 1 },
+          formData: { Start: 2, End: 0 },
           customValidate,
         });
 
+        // live validate does not run on initial render anymore
+        expect(node.querySelectorAll('#root_Start__error')).to.have.length(0);
+
+        // Change the End field to a larger value than Start field to remove customValidate raised errors.
+        let endInput = node.querySelector('#root_End');
+        act(() => {
+          fireEvent.change(endInput, {
+            target: { value: 1 },
+          });
+        });
         expect(node.querySelectorAll('#root_Start__error')).to.have.length(1);
         let errorMessageContent = node.querySelector('#root_Start__error .text-danger').textContent;
         expect(errorMessageContent).to.contain('Validate error: Test should be LE than End');
 
         // Change the End field to a larger value than Start field to remove customValidate raised errors.
-        const endInput = node.querySelector('#root_End');
         act(() => {
           fireEvent.change(endInput, {
             target: { value: 3 },
@@ -5111,6 +5202,30 @@ describe('Calling reset from ref object', () => {
     expect(input.getAttribute('value')).to.eq('');
     expect(checkbox.checked).to.eq(true);
   });
+
+  it('Reset button test with initialFormData', () => {
+    const schemaWithDefault = {
+      title: 'Test form',
+      type: 'string',
+    };
+    const formRef = createRef();
+    const props = {
+      ref: formRef,
+      initialFormData: 'foo',
+      schema: schemaWithDefault,
+    };
+    const { node } = createFormComponent(props);
+    const input = node.querySelector('input');
+    expect(formRef.current.reset).to.exist;
+    expect(input).to.exist;
+    expect(input.getAttribute('value')).to.eq(props.initialFormData);
+    fireEvent.change(input, { target: { value: 'Changed value' } });
+    expect(input.getAttribute('value')).to.eq('Changed value');
+    act(() => {
+      formRef.current.reset();
+    });
+    expect(input.getAttribute('value')).to.eq(props.initialFormData);
+  });
 });
 
 describe('validateForm()', () => {
@@ -5215,6 +5330,176 @@ describe('validateForm()', () => {
       expect(ref.current.validateForm()).to.eql(true);
     });
     errors = node.querySelectorAll('.error-detail');
+    expect(errors).to.have.lengthOf(0);
+  });
+});
+
+describe('setFieldValue()', () => {
+  let sandbox;
+  beforeEach(() => {
+    sandbox = createSandbox();
+  });
+  afterEach(() => {
+    sandbox.restore();
+  });
+  it('Sets root to value using ""', () => {
+    const ref = createRef();
+    const props = {
+      schema: {
+        type: 'string',
+      },
+      formData: {},
+      ref,
+    };
+    const { onChange, node } = createFormComponent(props);
+    // populate the input
+    act(() => {
+      ref.current.setFieldValue('', 'populated value');
+    });
+
+    sinon.assert.calledWithMatch(
+      onChange.lastCall,
+      {
+        formData: 'populated value',
+      },
+      'root',
+    );
+
+    expect(node.querySelector('input').value).to.eql('populated value');
+  });
+  it('Sets root to value using []', () => {
+    const ref = createRef();
+    const props = {
+      schema: {
+        type: 'string',
+      },
+      formData: {},
+      ref,
+    };
+    const { onChange, node } = createFormComponent(props);
+    // populate the input
+    act(() => {
+      ref.current.setFieldValue([], 'populated value');
+    });
+
+    sinon.assert.calledWithMatch(
+      onChange.lastCall,
+      {
+        formData: 'populated value',
+      },
+      'root',
+    );
+
+    expect(node.querySelector('input').value).to.eql('populated value');
+  });
+  it('Sets field to new value via dotted path', () => {
+    const ref = createRef();
+    const props = {
+      schema: {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'object',
+            required: ['input'],
+            properties: {
+              input: {
+                type: 'string',
+              },
+            },
+          },
+        },
+        required: ['foo'],
+      },
+      formData: {},
+      ref,
+      liveValidate: true,
+    };
+    const { onChange, node } = createFormComponent(props);
+    // trigger programmatic validation and make sure an error appears.
+    act(() => {
+      expect(ref.current.validateForm()).to.eql(false);
+    });
+
+    let errors = node.querySelectorAll('.error-detail');
+    expect(errors).to.have.lengthOf(1);
+    expect(errors[0].textContent).to.be.eql("must have required property 'input'");
+
+    // populate the input
+    act(() => {
+      ref.current.setFieldValue('foo.input', 'populated value');
+    });
+
+    sinon.assert.calledWithMatch(
+      onChange.lastCall,
+      {
+        formData: {
+          foo: {
+            input: 'populated value',
+          },
+        },
+      },
+      'root_foo_input',
+    );
+
+    // error should still be present.
+    errors = node.querySelectorAll('.error-detail');
+    // screen.debug();
+    // change formData and make sure the error disappears.
+    expect(errors).to.have.lengthOf(0);
+  });
+  it('Sets field to new value via field path list', () => {
+    const ref = createRef();
+    const props = {
+      schema: {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'object',
+            required: ['input'],
+            properties: {
+              input: {
+                type: 'string',
+              },
+            },
+          },
+        },
+        required: ['foo'],
+      },
+      formData: {},
+      ref,
+      liveValidate: true,
+    };
+    const { onChange, node } = createFormComponent(props);
+    // trigger programmatic validation and make sure an error appears.
+    act(() => {
+      expect(ref.current.validateForm()).to.eql(false);
+    });
+
+    let errors = node.querySelectorAll('.error-detail');
+    expect(errors).to.have.lengthOf(1);
+    expect(errors[0].textContent).to.be.eql("must have required property 'input'");
+
+    // populate the input
+    act(() => {
+      ref.current.setFieldValue(['foo', 'input'], 'populated value');
+    });
+
+    sinon.assert.calledWithMatch(
+      onChange.lastCall,
+      {
+        formData: {
+          foo: {
+            input: 'populated value',
+          },
+        },
+      },
+      'root_foo_input',
+    );
+
+    // error should still be present.
+    errors = node.querySelectorAll('.error-detail');
+    // screen.debug();
+    // change formData and make sure the error disappears.
     expect(errors).to.have.lengthOf(0);
   });
 });
@@ -5759,5 +6044,78 @@ describe('nameGenerator', () => {
 
     const activeInput = node.querySelector('#root_active');
     expect(activeInput.getAttribute('name')).eql('root[active]');
+  });
+});
+
+describe('initialFormData feature to prevent form reset', () => {
+  const schema = {
+    title: 'Reset Example',
+    properties: {
+      name: { type: 'string', title: 'Name' },
+    },
+  };
+  const data = { name: 'initial_id' };
+  /** This was adapted from the [example](https://playcode.io/2038613) provided in issue #391
+   */
+  const FormWrapper = ({ formData, initialFormData }) => {
+    const [isPending, setIsPending] = useState(false);
+
+    const handleSubmit = () => {
+      setIsPending(true);
+      setTimeout(() => setIsPending(false), 3000);
+    };
+
+    const props = {
+      ...(formData !== undefined && { formData }),
+    };
+
+    return (
+      <Form
+        initialFormData={initialFormData}
+        schema={schema}
+        validator={validator}
+        onSubmit={handleSubmit}
+        disabled={isPending}
+        {...props}
+      />
+    );
+  };
+  it('show that Form resets without initial data when it is controlled', () => {
+    const { container } = render(<FormWrapper formData={data} />);
+    let input = container.querySelector('input');
+    expect(input.value).to.eql(data.name);
+
+    fireEvent.change(input, {
+      target: { value: 'new value' },
+    });
+    input = container.querySelector('input');
+    expect(input.value).to.eql('new value');
+
+    const submit = container.querySelector('button');
+    act(() => {
+      fireEvent.click(submit);
+    });
+
+    input = container.querySelector('input');
+    expect(input.value).to.eql(data.name);
+  });
+  it('show that Form does not reset with initialFormData when it is uncontrolled', () => {
+    const { container } = render(<FormWrapper initialFormData={data} />);
+    let input = container.querySelector('input');
+    expect(input.value).to.eql(data.name);
+
+    fireEvent.change(input, {
+      target: { value: 'new value' },
+    });
+    input = container.querySelector('input');
+    expect(input.value).to.eql('new value');
+
+    const submit = container.querySelector('button');
+    act(() => {
+      fireEvent.click(submit);
+    });
+
+    input = container.querySelector('input');
+    expect(input.value).to.eql('new value');
   });
 });